move domain_type.rs
