+++ anchor/network/Cargo.toml
@@ -29,7 +29,7 @@ lighthouse_network = { workspace = true }
 quick-protobuf = "0.8.1"
 serde = { workspace = true }
 serde_json = "1.0.137"
-ssv_network_config = { workspace = true }
+ssv_types = { workspace = true }
 ssz_types = "0.10"
 subnet_tracker = { workspace = true }
 task_executor = { workspace = true }
